+++ hydrolib/core/io/ext/models.py
@@ -7,7 +7,7 @@ from hydrolib.core.basemodel import (
     DiskOnlyFileModel,
     validator_set_default_disk_only_file_model_when_none,
 )
-from hydrolib.core.io.bc.models import ForcingBase, ForcingData, ForcingModel
+from hydrolib.core.io.dflowfm.bc.models import ForcingBase, ForcingData, ForcingModel
 from hydrolib.core.io.ini.models import INIBasedModel, INIGeneral, INIModel
 from hydrolib.core.io.ini.serializer import SerializerConfig, write_ini
 from hydrolib.core.io.ini.util import (
