+++ tests/io/test_ext.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Optional
 import pytest
 from pydantic import ValidationError
 
-from hydrolib.core.io.bc.models import Constant, ForcingModel, RealTime
+from hydrolib.core.io.dflowfm.bc.models import Constant, ForcingModel, RealTime
 from hydrolib.core.io.ext.models import Boundary, ExtModel, Lateral
 from hydrolib.core.io.ini.models import INIBasedModel
